// yugabyted-ui/ui/src/api/src/models/ClusterResponse.ts
// tslint:disable
/**
 * Yugabyte Cloud
 * YugabyteDB as a Service
 *
 * The version of the OpenAPI document: v1
 * Contact: support@yugabyte.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// eslint-disable-next-line no-duplicate-imports
import type { ClusterData } from './ClusterData';


/**
 * 
 * @export
 * @interface ClusterResponse
 */
export interface ClusterResponse  {
  /**
   * 
   * @type {ClusterData}
   * @memberof ClusterResponse
   */
  data?: ClusterData;
}